double

+++ test/ExpressionCalculator/PenghitungTest.java
@@ -54,8 +54,8 @@ public class PenghitungTest {
         E.AddToken(O);
         E.AddToken(B1);
         Penghitung instance = new Penghitung();
-        int expResult = 2;
-        int result = instance.Calculate(E);
+        double expResult = 2;
+        double result = instance.Calculate(E);
         assertEquals(expResult, result, 2);
         // TODO review the generated test code and remove the default call to fail.
         //fail("The test case is a prototype.");
@@ -85,8 +85,8 @@ public class PenghitungTest {
         int b = 0;
         Operator o = new Operator("+");
         Penghitung instance = new Penghitung();
-        int expResult = 1;
-        int result = instance.CalculateAtom(a, b, o);
+        double expResult = 1;
+        double result = instance.CalculateAtom(a, b, o);
         assertEquals(expResult, result, 1);
         // TODO review the generated test code and remove the default call to fail.
         //fail("The test case is a prototype.");
@@ -107,8 +107,8 @@ public class PenghitungTest {
         E.AddToken(O);
         
         Penghitung instance = new Penghitung();
-        int expResult = 2;
-        int result = instance.CalculatePostfix(E);
+        double expResult = 2;
+        double result = instance.CalculatePostfix(E);
         assertEquals(expResult, result, 2);
         // TODO review the generated test code and remove the default call to fail.
         //fail("The test case is a prototype.");
